A_project/SMC_tidy/requirements.txt
# SMC Technical Analysis Tool Dependencies
# ========================================
# Use with conda environment: conda create -n smc_env python=3.11

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Data Fetching
akshare>=1.18.0
yfinance>=0.2.0

# SMC Analysis
smartmoneyconcepts>=0.0.26

# Visualization
plotly>=6.0.0

# Configuration
pyyaml>=6.0

# Excel Output
openpyxl>=3.1.0

# CLI Beautification
rich>=13.0.0

# Progress Bar
tqdm>=4.65.0

# Web Framework
nicegui>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Async Support
aiohttp>=3.8.0
httpx>=0.24.0

# Caching
diskcache>=5.6.0

# System Monitoring
psutil>=5.9.0

# Testing (dev)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
